refactor: DTO의 price, bargainPrice, discountedPrice 타입 Double로 변경

+++ MMarket/MMarket/Data/DTO/ProductDTO.swift
@@ -14,9 +14,9 @@ struct ProductDTO: Decodable {
     var description: String?
     let thumbnail: String
     let currency: String
-    let price: Int
-    let bargainPrice: Int
-    let discountedPrice: Int
+    let price: Double
+    let bargainPrice: Double
+    let discountedPrice: Double
     let stock: Int
     var images: [ProductImageDTO]?
     var vendor: VendorDTO?
